(edit) replace runtime data fetching with getServerSideProps in [accession_code] and [title] (previously [experiment]) and adjust codebase accordingly
